requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
pipenv==2023.10.3
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
whitenoise==6.5.0
